+++ wzz-parent/wzz-core/src/main/java/com/gmh/wzz/core/service/impl/WzzServiceImpl.java
@@ -42,7 +42,7 @@ public class WzzServiceImpl implements WzzService {
 	private String wzz_ftp_userName = "www";
 	private String wzz_ftp_password = "a85e0d3e6";
 	private String wzz_ftp_tmpfile_path = "/mydata/tmp";
-	private int wzz_wifi_search_scope = 500;
+	private Float wzz_wifi_search_scope = 500f;
 
 	private String SMSServerHost;
 	private String SMSServerPort;
@@ -137,11 +137,11 @@ public class WzzServiceImpl implements WzzService {
 		this.wzz_ftp_tmpfile_path = wzz_ftp_tmpfile_path;
 	}
 
-	public int getWzz_wifi_search_scope() {
+	public Float getWzz_wifi_search_scope() {
 		return wzz_wifi_search_scope;
 	}
 
-	public void setWzz_wifi_search_scope(int wzz_wifi_search_scope) {
+	public void setWzz_wifi_search_scope(Float wzz_wifi_search_scope) {
 		this.wzz_wifi_search_scope = wzz_wifi_search_scope;
 	}
